requirements.txt
flask
pyyaml
llama-index
llama-index-llms-openai-like
llama-index-readers-file
llama-index-embeddings-huggingface
sentence-transformers
pypdf  # llama-index读取pdf需要
requests
ipython # 提升命令行客户端的体验
PyPDF2==3.0.1
pdfplumber==0.10.3
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.2.0
beautifulsoup4==4.12.3
selenium==4.18.1
openai>=1.0.0
transformers>=4.30.0
json-repair>=0.1.0
aiohttp>=3.8.0
tqdm>=4.65.0
asyncio>=3.4.3
typing-extensions>=4.5.0
flask_cors